use %zu rather than casting to long

+++ src/libmongoc/src/mongoc/mongoc-stream-tls-openssl.c
@@ -342,16 +342,16 @@ _mongoc_stream_tls_openssl_writev (mongoc_stream_t *stream,
             child_ret =
                _mongoc_stream_tls_openssl_write (tls, to_write, to_write_len);
             if (bson_cmp_not_equal_su (child_ret, to_write_len)) {
-               TRACE ("Got child_ret: %ld while to_write_len is: %ld",
-                      (long)child_ret,
-                      (long)to_write_len);
+               TRACE ("Got child_ret: %zu while to_write_len is: %zu",
+                      child_ret,
+                      to_write_len);
             }
 
             if (child_ret < 0) {
-               TRACE ("Returning what I had (%ld) as apposed to the error "
-                      "(%ld, errno:%d)",
-                      (long)ret,
-                      (long)child_ret,
+               TRACE ("Returning what I had (%zu) as apposed to the error "
+                      "(%zu, errno:%d)",
+                      ret,
+                      child_ret,
                       errno);
                RETURN (ret);
             }
